+++ CHANGELOG.md
@@ -42,6 +42,7 @@ Changes since v2.4
 - Improved table widget
 - Hide language switcher when only one language configured
 - PDFs are now downloaded as \<application-id\>.pdf instead of pdf.pdf
+- Improved table performance: added a "show all rows" button for long tables
 
 ### Fixes
 - Entitlement API